requirements.txt
google-generativeai
langchain
langchain-community
langchain-google-genai
qdrant-client
uuid
fastapi
uvicorn
python-dotenv